requirements.txt
python-dotenv
mysql.connector
datetime